// src/network/openapi/v1.d.ts
export interface components {
    schemas: {
// ...
        FoodItemDto: {
            /** Format: int32 */
            id: number;
            name: string | null;
            description: string | null;
            subtext: string | null;
            imageUrl: string | null;
            nutritionInfo: components["schemas"]["FoodItemNutritionInfoDto"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}